+++ tests/lib/Acceptance/Relations/MorphTo/QueryTest.php
@@ -22,6 +22,8 @@ namespace LaravelJsonApi\Eloquent\Tests\Acceptance\Relations\MorphTo;
 use App\Models\Image;
 use App\Models\Post;
 use App\Models\User;
+use App\Schemas\PostSchema;
+use App\Schemas\UserSchema;
 
 class QueryTest extends TestCase
 {
@@ -92,6 +94,39 @@ class QueryTest extends TestCase
         $actual = $this->repository
             ->queryToOne($image, 'imageable')
             ->with($path)
+            ->first();
+
+        $this->assertTrue($image->imageable->is($actual));
+        $this->assertTrue($actual->relationLoaded($relation));
+    }
+
+    /**
+     * @return array
+     */
+    public function defaultEagerLoadProvider(): array
+    {
+        return [
+            'post' => [PostSchema::class, Post::class, 'user'],
+            'user' => [UserSchema::class, User::class, 'country'],
+        ];
+    }
+
+    /**
+     * @param string $schemaClass
+     * @param string $modelClass
+     * @param string $relation
+     * @dataProvider defaultEagerLoadProvider
+     */
+    public function testWithDefaultEagerLoading(string $schemaClass, string $modelClass, string $relation): void
+    {
+        $this->createSchemaWithDefaultEagerLoading($schemaClass, $relation);
+
+        $image = Image::factory()
+            ->for($modelClass::factory(), 'imageable')
+            ->create();
+
+        $actual = $this->repository
+            ->queryToOne($image, 'imageable')
             ->first();
 
         $this->assertTrue($image->imageable->is($actual));